requirements.txt
# Core dependencies for ComfyUI nodes
requests>=2.25.0

# Model loading dependencies
huggingface_hub>=0.20.0
safetensors>=0.4.0

# Video processing dependencies
scenedetect>=0.6.0
opencv-python>=4.5.0

# Loop extraction dependencies
imagehash>=4.2.0
imageio>=2.25.0
imageio-ffmpeg>=0.4.9  # Bundles FFmpeg binary - no system install needed
Pillow>=8.0.0
numpy
scipy
fire
matplotlib
transformers
torch
torchvision  # Required for BiRefNet preprocessing

# Note: imageio-ffmpeg includes a bundled FFmpeg binary, so no system
# installation of FFmpeg is required. This makes setup much easier!
